requirements.txt
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
colorama==0.4.5
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.7.10
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.3
langdetect==1.0.9
loguru==0.6.0
numpy==1.23.1
peewee==3.15.1
pybind11==2.10.0
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
telebot-calendar==1.2
urllib3==1.26.10
win32-setctime==1.1.0
